# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Otacon Phone Automation API",
        "description": "REST API for controlling Android phones connected to a Raspberry Pi.",
        "license": {
            "name": "MIT"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://otacon-pi.<tailnet>.ts.net:8080",
            "description": "Default Tailscale address"
        }
    ],
    "paths": {
...
        "/api/sims": {
            "get": {
                "tags": [
                    "SIM"
                ],
                "operationId": "listEsimProfiles",
                "parameters": [
                    {
                        "name": "all",
                        "in": "query",
                        "description": "Include historical (stale) physical SIM records",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/EsimProfile"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/sims/defaults": {
            "get": {
                "tags": [
                    "SIM"
                ],
                "operationId": "getEsimDefaults",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EsimDefaults"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "SIM"
                ],
                "operationId": "setEsimDefaults",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetDefaultsBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/sims/delete": {
            "post": {
                "tags": [
                    "SIM"
                ],
                "operationId": "deleteEsimProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DeleteBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/sims/enable": {
            "post": {
                "tags": [
                    "SIM"
                ],
                "operationId": "enableEsimProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EnableBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/sims/install": {
            "post": {
                "tags": [
                    "SIM"
                ],
                "operationId": "installEsimProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InstallBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/sims/switch": {
            "post": {
                "tags": [
                    "SIM"
                ],
                "operationId": "switchEsimProfile",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SwitchBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DeleteBody": {
                "type": "object",
                "required": [
                    "subId"
                ],
                "properties": {
                    "subId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "EnableBody": {
                "type": "object",
                "required": [
                    "subId",
                    "enabled"
                ],
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    },
                    "subId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "EsimDefaults": {
                "type": "object",
                "properties": {
                    "dataSubId": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    },
                    "smsSubId": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    },
                    "voiceSubId": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    }
                }
            },
            "EsimProfile": {
                "type": "object",
                "required": [
                    "subId",
                    "iccid",
                    "carrier",
                    "slot",
                    "embedded",
                    "enabled",
                    "status",
                    "isDefault"
                ],
                "properties": {
                    "carrier": {
                        "type": "string"
                    },
                    "embedded": {
                        "type": "boolean"
                    },
                    "enabled": {
                        "type": "boolean",
                        "description": "Currently active on a SIM slot (slot >= 0 && apps enabled)"
                    },
                    "historical": {
                        "type": "boolean",
                        "description": "True for stale physical SIM records (only present when ?all=true)"
                    },
                    "iccid": {
                        "type": "string"
                    },
                    "isDefault": {
                        "type": "boolean"
                    },
                    "slot": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "status": {
                        "type": "string",
                        "description": "Status string: \"active\" (enabled, on a slot), \"disabled\" (installed\nbut not active), or \"historical\" (stale subscription record from a\npreviously-inserted physical SIM, only shown with ?all=true)"
                    },
                    "subId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "InstallBody": {
                "type": "object",
                "required": [
                    "activationCode"
                ],
                "properties": {
                    "activationCode": {
                        "type": "string"
                    }
                }
            },
...
            "SetDefaultsBody": {
                "type": "object",
                "properties": {
                    "dataSubId": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    },
                    "smsSubId": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    },
                    "voiceSubId": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int64"
                    }
                }
            },
...
            "SwitchBody": {
                "type": "object",
                "required": [
                    "subId"
                ],
                "properties": {
                    "subId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
        }
    }
}